third_party/blink/web_tests/platform/generic/fast/css/variables/custom-properties-in-object-model-expected.txt

PASS style.getPropertyValue("--important") is "green"
PASS style.getPropertyValue("--not-important") is "red"
PASS style.getPropertyValue("color") is "var(--important)"
PASS style.getPropertyPriority("--important") is "important"
PASS style.getPropertyPriority("--not-important") is ""
PASS style.getPropertyValue("--foo") is "papayawhip"
PASS style.getPropertyPriority("--foo") is ""
PASS style.getPropertyPriority("--important-foo") is "important"
PASS style.getPropertyValue("--important-foo") is ""
PASS style.removeProperty("--foo") is "papayawhip"
PASS style.getPropertyValue("--foo") is ""
PASS computedStyle.setProperty("--error", "") threw exception NoModificationAllowedError: Failed to execute 'setProperty' on 'CSSStyleDeclaration': These styles are computed, and therefore the '--error' property is read-only..
PASS cssText is "#test2 { --variable:value; }"
PASS successfullyParsed is true

TEST COMPLETE
